Improvement to azure portal command.
We now try to infer the environment the user is likely to prefer, when
they omitted the -e argument to the command. With this change, if -e is
omitted, we open the portal for the environment associated with the
user's subscription that they are currently signed into. Other cases (
where -e is specified, user is not signed in etc.) remain the same as
before.